cart backend changes added

+++ restaurant-backend/src/main/java/com/app/restaurant/controller/EmployeeController.java
@@ -31,7 +31,7 @@ public class EmployeeController {
 
     //build get employee by id REST API
     @GetMapping("{id}")
-    public ResponseEntity<Employee> getEmployeeById(@PathVariable long id){
+    public ResponseEntity<Employee> getEmployeeById(@PathVariable int id){
         Employee employee =employeeRepo.findById(id).orElseThrow(() -> new ResourceNotFoundException("Employee not exist with id " +id));
 
         return ResponseEntity.ok(employee);
@@ -39,7 +39,7 @@ public class EmployeeController {
 
     //build update employee REST API
     @PutMapping("{id}")
-    public ResponseEntity<Employee> updateEmployee(@PathVariable long id,@RequestBody Employee employeeDetails){
+    public ResponseEntity<Employee> updateEmployee(@PathVariable int id,@RequestBody Employee employeeDetails){
         Employee updateEmployee =employeeRepo.findById(id).orElseThrow(() -> new ResourceNotFoundException("Employee not exist with id " +id));
 
         updateEmployee.setFirstName(employeeDetails.getFirstName());
@@ -57,7 +57,7 @@ public class EmployeeController {
 
     // build delete employee REST API
     @DeleteMapping("{id}")
-    public ResponseEntity<HttpStatus> deleteEmployee(@PathVariable long id){
+    public ResponseEntity<HttpStatus> deleteEmployee(@PathVariable int id){
         Employee employee =employeeRepo.findById(id).orElseThrow(() -> new ResourceNotFoundException("Employee not exist with id " +id));
 
         employeeRepo.delete(employee);
