+++ restaurant-backend/src/main/java/com/app/restaurant/model/Employee.java
@@ -18,7 +18,7 @@ public class Employee {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long employee_id;
+    private int employee_id;
 
     @Column(name="first_name" , nullable = false)
     private String firstName;
