+++ restaurant-backend/src/main/java/com/app/restaurant/repository/MealRepository.java
@@ -3,5 +3,5 @@ package com.app.restaurant.repository;
 import com.app.restaurant.model.Meal;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MealRepository extends JpaRepository<Meal, Long> {
+public interface MealRepository extends JpaRepository<Meal, Integer> {
 }
